# frontend/types/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Swell Travel Agent API",
  "description": "Surf-forecast-driven travel discovery. All state-changing requests need the session cookie and an `X-CSRF-Token` header matching the `sta_csrf` cookie (GET `/api/auth/csrf` first).",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/users/me/profile": {
   "patch": {
    "tags": [
     "account"
    ],
    "summary": "Update Profile",
    "operationId": "update_profile_api_users_me_profile_patch",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProfileIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/users/me/notification-preferences": {
   "patch": {
    "tags": [
     "account"
    ],
    "summary": "Update Notification Prefs",
    "operationId": "update_notification_prefs_api_users_me_notification_preferences_patch",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/NotificationPrefsIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/users/me/phone/start": {
   "post": {
    "tags": [
     "account"
    ],
    "summary": "Start Phone",
    "operationId": "start_phone_api_users_me_phone_start_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PhoneStartIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Message"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/users/me/phone/confirm": {
   "post": {
    "tags": [
     "account"
    ],
    "summary": "Confirm Phone",
    "operationId": "confirm_phone_api_users_me_phone_confirm_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PhoneConfirmIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/users/me/airports": {
   "get": {
    "tags": [
     "account"
    ],
    "summary": "List Airports",
    "operationId": "list_airports_api_users_me_airports_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/UserAirportOut"
         },
         "type": "array",
         "title": "Response List Airports Api Users Me Airports Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "account"
    ],
    "summary": "Add Airport",
    "operationId": "add_airport_api_users_me_airports_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserAirportIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/UserAirportOut"
         },
         "type": "array",
         "title": "Response Add Airport Api Users Me Airports Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/users/me/airports/{airport_id}": {
   "delete": {
    "tags": [
     "account"
    ],
    "summary": "Remove Airport",
    "operationId": "remove_airport_api_users_me_airports__airport_id__delete",
    "parameters": [
     {
      "name": "airport_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Airport Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/UserAirportOut"
         },
         "title": "Response Remove Airport Api Users Me Airports  Airport Id  Delete"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/users/me": {
   "delete": {
    "tags": [
     "account"
    ],
    "summary": "Delete Account",
    "operationId": "delete_account_api_users_me_delete",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DeleteAccountIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Message"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/airports": {
   "get": {
    "tags": [
     "spots"
    ],
    "summary": "Search Airports",
    "operationId": "search_airports_api_airports_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "maxLength": 60,
       "default": "",
       "title": "Q"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 20,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/AirportOut"
         },
         "title": "Response Search Airports Api Airports Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "AirportOut": {
    "properties": {
     "iata": {
      "type": "string",
      "title": "Iata"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "city": {
      "type": "string",
      "title": "City"
     },
     "country": {
      "type": "string",
      "title": "Country"
     },
     "country_code": {
      "type": "string",
      "title": "Country Code"
     },
     "latitude": {
      "type": "number",
      "title": "Latitude"
     },
     "longitude": {
      "type": "number",
      "title": "Longitude"
     },
     "timezone": {
      "type": "string",
      "title": "Timezone"
     }
    },
    "type": "object",
    "required": [
     "iata",
     "name",
     "city",
     "country",
     "country_code",
     "latitude",
     "longitude",
     "timezone"
    ],
    "title": "AirportOut"
   },
...
   "BreakType": {
    "type": "string",
    "enum": [
     "reef",
     "point",
     "beach",
     "rivermouth",
     "slab"
    ],
    "title": "BreakType"
   },
...
   "DeleteAccountIn": {
    "properties": {
     "password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 1,
      "title": "Password"
     },
     "confirm": {
      "type": "string",
      "const": "DELETE",
      "title": "Confirm"
     }
    },
    "type": "object",
    "required": [
     "password",
     "confirm"
    ],
    "title": "DeleteAccountIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Message": {
    "properties": {
     "message": {
      "type": "string",
      "title": "Message"
     }
    },
    "type": "object",
    "required": [
     "message"
    ],
    "title": "Message"
   },
...
   "NotificationPrefsIn": {
    "properties": {
     "email_enabled": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Email Enabled"
     },
     "sms_enabled": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sms Enabled"
     },
     "all_paused": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "All Paused"
     },
     "max_alerts_per_day": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 100.0,
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Alerts Per Day"
     }
    },
    "type": "object",
    "title": "NotificationPrefsIn"
   },
   "NotificationPrefsOut": {
    "properties": {
     "email_enabled": {
      "type": "boolean",
      "title": "Email Enabled"
     },
     "sms_enabled": {
      "type": "boolean",
      "title": "Sms Enabled"
     },
     "all_paused": {
      "type": "boolean",
      "title": "All Paused"
     },
     "phone_number": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone Number"
     },
     "phone_verified": {
      "type": "boolean",
      "title": "Phone Verified"
     },
     "sms_ready": {
      "type": "boolean",
      "title": "Sms Ready"
     },
     "sms_opted_out": {
      "type": "boolean",
      "title": "Sms Opted Out"
     },
     "max_alerts_per_day": {
      "type": "integer",
      "title": "Max Alerts Per Day"
     },
     "sms_service_available": {
      "type": "boolean",
      "title": "Sms Service Available"
     }
    },
    "type": "object",
    "required": [
     "email_enabled",
     "sms_enabled",
     "all_paused",
     "phone_number",
     "phone_verified",
     "sms_ready",
     "sms_opted_out",
     "max_alerts_per_day",
     "sms_service_available"
    ],
    "title": "NotificationPrefsOut"
   },
...
   "PhoneConfirmIn": {
    "properties": {
     "code": {
      "type": "string",
      "maxLength": 10,
      "minLength": 4,
      "title": "Code"
     }
    },
    "type": "object",
    "required": [
     "code"
    ],
    "title": "PhoneConfirmIn"
   },
   "PhoneStartIn": {
    "properties": {
     "phone_number": {
      "type": "string",
      "maxLength": 24,
      "minLength": 7,
      "title": "Phone Number"
     }
    },
    "type": "object",
    "required": [
     "phone_number"
    ],
    "title": "PhoneStartIn"
   },
   "ProfileIn": {
    "properties": {
     "full_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 120,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Full Name"
     },
     "home_airport": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Home Airport"
     },
     "home_city": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 120
       },
       {
        "type": "null"
       }
      ],
      "title": "Home City"
     },
     "preferred_wave_min_ft": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 80.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Preferred Wave Min Ft"
     },
     "preferred_wave_max_ft": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 80.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Preferred Wave Max Ft"
     },
     "preferred_break_type": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/BreakType"
       },
       {
        "type": "null"
       }
      ]
     },
     "experience_level": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/SkillLevel"
       },
       {
        "type": "null"
       }
      ]
     },
     "units": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "ft",
         "m"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Units"
     },
     "currency": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Currency"
     }
    },
    "type": "object",
    "title": "ProfileIn"
   },
   "ProfileOut": {
    "properties": {
     "home_airport": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Home Airport"
     },
     "home_city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Home City"
     },
     "preferred_wave_min_ft": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Preferred Wave Min Ft"
     },
     "preferred_wave_max_ft": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Preferred Wave Max Ft"
     },
     "preferred_break_type": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/BreakType"
       },
       {
        "type": "null"
       }
      ]
     },
     "experience_level": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/SkillLevel"
       },
       {
        "type": "null"
       }
      ]
     },
     "units": {
      "type": "string",
      "title": "Units"
     },
     "currency": {
      "type": "string",
      "title": "Currency"
     }
    },
    "type": "object",
    "required": [
     "home_airport",
     "home_city",
     "preferred_wave_min_ft",
     "preferred_wave_max_ft",
     "preferred_break_type",
     "experience_level",
     "units",
     "currency"
    ],
    "title": "ProfileOut"
   },
...
   "SkillLevel": {
    "type": "string",
    "enum": [
     "beginner",
     "intermediate",
     "advanced",
     "expert"
    ],
    "title": "SkillLevel"
   },
...
   "UserAirportIn": {
    "properties": {
     "airport_iata": {
      "type": "string",
      "title": "Airport Iata"
     },
     "label": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 60
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     },
     "is_home": {
      "type": "boolean",
      "title": "Is Home",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "airport_iata"
    ],
    "title": "UserAirportIn"
   },
   "UserAirportOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "airport_iata": {
      "type": "string",
      "title": "Airport Iata"
     },
     "label": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     },
     "is_home": {
      "type": "boolean",
      "title": "Is Home"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "city": {
      "type": "string",
      "title": "City"
     },
     "country": {
      "type": "string",
      "title": "Country"
     }
    },
    "type": "object",
    "required": [
     "id",
     "airport_iata",
     "label",
     "is_home",
     "name",
     "city",
     "country"
    ],
    "title": "UserAirportOut"
   },
   "UserOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "full_name": {
      "type": "string",
      "title": "Full Name"
     },
     "email_verified": {
      "type": "boolean",
      "title": "Email Verified"
     },
     "is_admin": {
      "type": "boolean",
      "title": "Is Admin"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "profile": {
      "$ref": "#/components/schemas/ProfileOut"
     },
     "notification_preferences": {
      "$ref": "#/components/schemas/NotificationPrefsOut"
     },
     "airports": {
      "items": {
       "$ref": "#/components/schemas/UserAirportOut"
      },
      "type": "array",
      "title": "Airports"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "full_name",
     "email_verified",
     "is_admin",
     "created_at",
     "profile",
     "notification_preferences",
     "airports"
    ],
    "title": "UserOut"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}